ref(utils): refactor utils into utils subdir
- split util fns into `descriptors` and `common`
